Split models (Sonnet for reflect, Opus for learn/advise) and rename Ask to Advise
- Fix Bedrock thinking config: move effort to output_config (was incorrectly nested inside thinking object)
- Add supportsThinking() guard for non-Claude models
- Make NewClient accept a model parameter with ModelOpus/ModelSonnet constants
- Dream pipeline takes two LLMs: Sonnet for reflect (fast extraction), Opus for learn (synthesis)
- Rename Ask to Advise across CLI, MCP tool, and shade package
- Filter whitespace-only observations from reflect phase
